Navigation Burger fixes to text, scaling, images, icons.
Icons and Images were extremely faded. This fixes that. Along with this text was relatively small so increased text size. Also needed to ensure that the changes did not affect anything outside the navigation burger.

By adding this it will make the icons and images visable, the text to be more readable, and the changes to remain in the navigation burger.