Updated CSP
The Content Security Policy has been updated to include the product's current resources being used. This should cut down a lot of false positives that are shown in CSP reporting. The Content Security Policy has been flagged as vulnerable in scans.

To make sure the CSP works for the product, we should verify that the product is actually reporting CSP warnings and see it block unrecognised resources (perhaps when the Contact page has been implemented)