Readability and Bubble/Dot changes
Changes the .bubbles to .dots. Changed the text for titles to be smaller to fit. Effectively makes the entire process easier to read for mobile users. Some issues remain with needing to refresh to see changes, however a phone user should not require going from web to mobile view constantly.