StatisticPage_UI_DEV_VictorXiao_#PR3
[Statistics page UI-development] https://trello.com/c/2OEiIsAi/541-statistics-page-ui-development-max-2

Created Barplot for Popularity Range.
Created filter drop list for years.
Created filter drop list for trimesters.
Fixed the issue that search function, now allowing users to search regardless of the case used.
Fixed the issue that the plots are not centered.
Fixed the issue that the search bar doesn't have full width of below table.
Extended the distance between body and footer.